(BSR)[API] chore: move cgr exception
