--- api/src/pcapi/connectors/cgr/exceptions.py
@@ -1,3 +1,0 @@
-# TODO(yacine) move this file to core.externalBookings.cgr
-class CGRAPIException(Exception):
-    pass
