+++ api/src/pcapi/core/external_bookings/cgr/exceptions.py
@@ -1,0 +1,2 @@
+class CGRAPIException(Exception):
+    pass
